backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "clerk-backend-api>=3.0.5",
    "fastapi>=0.116.1",
    "httpx>=0.28.1",
    "pydantic[email]>=2.11.7",
    "pydantic-settings>=2.10.1",
    "pymongo>=4.13.2",
    "uvicorn>=0.35.0",
    "pyjwt>=2.10.1",
    "slowapi>=0.1.9",
]
